+++ packages/tests/web-test-runner.config.js
@@ -151,7 +151,7 @@ export default {
   files: [
     '../lit-html/development/**/*_test.(js|html)',
     '../lit-element/development/**/*_test.(js|html)',
-    '../updating-element/development/**/*_test.(js|html)',
+    '../reactive-element/development/**/*_test.(js|html)',
   ],
   nodeResolve: true,
   concurrency: 6, // default cores / 2