+++ rollup-common.js
@@ -73,7 +73,7 @@ const stableProperties = {
   _$getTemplate: 'F',
   // lit-html: TemplateInstance (used by platform-support)
   _$template: 'G',
-  // updating-element: UpdatingElement (used by platform-support)
+  // reactive-element: ReactiveElement (used by platform-support)
   _$didUpdate: 'H',
   // lit-element: LitElement
   _$renderOptions: 'I',
@@ -87,7 +87,7 @@ const stableProperties = {
   _$setValue: 'M',
   // platform-support: LitElement (added by platform-support)
   _$handlesPrepareStyles: 'N',
-  // lit-element: UpdatingElement (used bby ssr-support)
+  // lit-element: ReactiveElement (used bby ssr-support)
   _$attributeToProperty: 'O',
   // lit-html: ChildPart, AttributePart, TemplateInstance, Directive (accessed by
   // disconnectable-directive)